# parseArgs: replace Sprintf with string concat